flights/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
Flask-Cors==5.0.0
gunicorn==21.2.0
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
pyroscope-io==0.8.7